+++ src/commands/form/FormDefinitionExportCommand.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs';
 import * as vscode from 'vscode';
 
-import { askFile, openPreview } from '../../utils/vsCodeHelpers';
+import { askChosenItems, askFile, openPreview } from '../../utils/vsCodeHelpers';
 import { FormTreeItem, FormsTreeItem } from '../../models/ISCTreeItem';
 import { PathProposer } from '../../services/PathProposer';
 import { ISCClient } from '../../services/ISCClient';
@@ -11,10 +11,10 @@ export class FormDefinitionExportCommand {
 
     async execute(node: FormsTreeItem | FormTreeItem): Promise<void> {
         console.log("> FormDefinitionExportCommand.execute");
-
+        const client = new ISCClient(node.tenantId, node.tenantName);
         let exportFile = "",
             prompt = "",
-            filters: string | undefined = undefined,
+            data,
             successfullMessage = ""
         if (node.hasOwnProperty("id")) {
             // FormTreeItem
@@ -24,7 +24,8 @@ export class FormDefinitionExportCommand {
                 node.label as string
             )
             prompt = `Enter the file to save ${node.label} to`
-            filters = `name eq "${node.label}"`
+            const filters = `name eq "${node.label}"`
+            data = await client.exportForms(filters)
             successfullMessage = `Successfully exported ${node.label} from ${node.tenantDisplayName}`
         } else {
             //FormsTreeItem
@@ -35,6 +36,20 @@ export class FormDefinitionExportCommand {
             prompt = `Enter the file to save forms from ${node.tenantDisplayName} to`
             successfullMessage = `Successfully exported forms from ${node.tenantDisplayName}`
 
+            const forms = (await client.exportForms()).map(x => ({
+                ...x,
+                id: x.object.id,
+                name: x.object.name,
+                description: x.object.description
+            }))
+            data = await askChosenItems("Exporting forms", "Forms", forms, x => {
+                const { id, name, description, ...rest } = x;
+                return rest;
+            })
+            if (data === undefined) {
+                return
+            }
+
         }
 
         const target = await askFile(
@@ -44,8 +59,6 @@ export class FormDefinitionExportCommand {
             return;
         }
 
-        const client = new ISCClient(node.tenantId, node.tenantName);
-        let data = await client.exportForms(filters)
         // Cleaning "data" by removing usedBy
         data = data.map(item => ({
             ...item,